Gradually moving metadata to class
